Move color sensor to its own subsystem
