+++ robot.py
@@ -29,11 +29,11 @@ RIGHT_HAND = wpilib._wpilib.XboxController.Hand.kRightHand
 
 class Robot(wpilib.TimedRobot):
     def robotInit(self):
-        #Controllers
+        # Controllers 
         self.driver = wpilib.XboxController(0)
         self.operator = wpilib.XboxController(1)
 
-        #Motors
+        # Motors
         self.left_motor_1 = rev.CANSparkMax(robotmap.LEFT_LEADER_ID, rev.MotorType.kBrushed)
         self.left_motor_2 = rev.CANSparkMax(robotmap.LEFT_FOLLOWER_ID, rev.MotorType.kBrushed)
         self.right_motor_1 = rev.CANSparkMax(robotmap.RIGHT_LEADER_ID, rev.MotorType.kBrushed)
@@ -44,14 +44,12 @@ class Robot(wpilib.TimedRobot):
         self.right_motor_1.setClosedLoopRampRate(1.0)
         self.right_motor_2.setClosedLoopRampRate(1.0)
         
+        # Group motors
         self.left_side = wpilib.SpeedControllerGroup(self.left_motor_1, self.left_motor_2)
         self.right_side = wpilib.SpeedControllerGroup(self.right_motor_1, self.right_motor_2)
         
         #Drivetrain
         self.drivetrain = wpilib.drive.DifferentialDrive(self.left_side, self.right_side)
-
-        #TODO: Add subsystems and sensors as the code is written
-        #TODO: SmartDashboard
         
         #Pneumatics
         self.colorPiston = pneumatic_system(wpilib.DoubleSolenoid(0, robotmap.COLOR_SENSOR_EXTEND, robotmap.COLOR_SENSOR_RETRACT))
@@ -63,19 +61,8 @@ class Robot(wpilib.TimedRobot):
 
         # Color Sensor
         self.colorSensor = color_sensor()
-        self.colorSensorMotor = rev.CANSparkMax(robotmap.COLOR_SENSOR_MOTOR, rev.MotorType.kBrushed)
-       
-        self.stopColorMap = {"R":"B", "Y":"G", "B":"R", "G":"Y"}
-        
-        self.gameData = ""
-        
-        self.setupColorSensor()
-
-        
-        
 
         #Shooter
-    
         self.shooter = shooter(robotmap.LOADER, robotmap.SHOOTER)
 
         # Gyro
@@ -86,24 +73,6 @@ class Robot(wpilib.TimedRobot):
         self.sd = NetworkTables.getTable('VISION')
         
         
-
-
-    def setupColorSensor(self):
-        self.colorMatch = ColorMatch()
-        
-        self.blue = wpilib._wpilib.Color(0.143, 0.427, 0.429)
-        self.green = wpilib._wpilib.Color(0.197, 0.561, 0.240)
-        self.red = wpilib._wpilib.Color(0.561, 0.232, 0.144)
-        self.yellow = wpilib._wpilib.Color(0.361, 0.524, 0.133)
-
-        self.colorMap = {"B":self.blue, "G":self.green, "R":self.red, "Y":self.yellow}
-        
-        self.colorMatch.addColorMatch(self.blue)
-        self.colorMatch.addColorMatch(self.green)
-        self.colorMatch.addColorMatch(self.red)
-        self.colorMatch.addColorMatch(self.yellow)
-
-
     def robotPeriodic(self):
         return
     
@@ -117,26 +86,11 @@ class Robot(wpilib.TimedRobot):
 
         self.Aimer.reset()
 
-        #self.Aimer.setaim(self.Aimer.getAngle())
         self.turned180 = False
-
         self.setTarget = False
 
-        #
+        #Reset aimer
         self.Aimer.setaim(self.Aimer.getAngle())
-
-    """
-    def rotateToPoint(self):
-        val = (self.Aimer.getAngle()-self.Aimer.getsetpoint())
-        #print(val)
-        if val > 6 or val < -6:
-            self.drivetrain.arcadeDrive(0, 0.7)
-        else:
-            self.drivetrain.arcadeDrive(0,0)
-            return True
-
-    """
-
 
     def autonomousPeriodic(self):
         
@@ -165,52 +119,10 @@ class Robot(wpilib.TimedRobot):
         self.shooter.setShooterSpeed(lspeed, robotmap.SHOOTER_RPM)
 
 
-        
-        #Move forward for 1 second
-        #
-        #else:
-        #Rotate 180, then rotate to target
-        #
-        
-
-        #if(abs(amt) < 1):
-        #    self.autonTimer.start()
-            
-        
-        #if amt == 0:
-        #    self.turned180 = True
-        #    if not self.setTarget:
-        #        self.Aimer.setaim(self.Aimer.getAngle() + self.sd.getNumber("ANGLE", 0))
-        #        self.setTarget = True
-        #    else:
-        #        self.Aimer.setaim(self.Aimer.getAngle())
-        #        
-                
-                
-            
-                
-            
-                
-                    
-            
-        """if self.rotateToPoint():
-            if not self.foundTarget:
-                self.Aimer.setaim(self.Aimer.getAngle() + self.sd.getNumber("ANGLE", 0))
-                print(self.sd.getNumber("ANGLE", 0))
-                self.foundTarget = True
-            
-            #self.Aimer.setaim(self.Aimer.getAngle())
-            #self.foundTarget = True
-        """
-        #if self.foundTarget:
-        #    self.Aimer.setaim(self.Aimer.getAngle())
-        #    self.shooter.setShooterSpeed(robotmap.LOADER_SPEED, robotmap.SHOOTER_RPM)
-
     def teleopInit(self):
         NetworkTables.initialize()
         self.sd = NetworkTables.getTable('VISION')
-       # self.sd = NetworkTables.getTable('VISION')
-        self.gameData = ""
+        
         self.goal = ""
         self.turnedAmount = 0
 
@@ -235,29 +147,14 @@ class Robot(wpilib.TimedRobot):
         self.hasTurnedWheel = False
 
 
-    def debugColorSensor(self, color=None):
-        if color is not None:
-            color = self.colorSensor.getColor()
-        red = color.red
-        blue = color.blue
-        green = color.green
-        # TODO: Use better debugging tools
-        print("Red: {} Green: {} Blue: {} ".format(red, green, blue))
-
-
-    def checkGameData(self):
-        gd = wpilib.DriverStation.getInstance().getGameSpecificMessage()
-        if(gd != None and not self.searchForColor):
-            self.gameData = gd
+    
 
 
     def colorPistonUpdate(self):
         if self.operator.getPOV() == 180:
             self.colorPiston.extend()
-            #self.colorArmIsExtended = True
         elif self.operator.getPOV() == 0:
             self.colorPiston.retract()
-            #self.colorArmIsExtended = False
 
 
     def climberPistonUpdate(self):  
@@ -284,88 +181,17 @@ class Robot(wpilib.TimedRobot):
         else:
             self.climberWinchMotor1.set(0)
             self.climberWinchMotor2.set(0)
-
-
-    def turnWheelInit(self):
-        self.turnedAmount = 8
-        self.currentColor = None
-        self.lastColor = None
-        self.startColor = self.stopColorMap[self.colorSensor.getColorName(self.colorSensor.getColor())]
-        self.lastColor = self.startColor
-        self.turnWheel = True
-
     
-    def turnWheelCycle(self):
-        #self.debugColorSensor()
-        self.colorSensorMotor.set(0.3)
-        self.currentColor = self.colorSensor.getColorName(self.colorSensor.getColor())
-        
-        if self.currentColor != self.lastColor:            
-            if self.currentColor == self.startColor:
-                self.turnedAmount -= 1           
-            if self.turnedAmount == 0:
-                self.colorSensorMotor.set(0)
-                self.turnWheel = False
-                self.hasTurnedWheel = True
-
-        self.lastColor = self.currentColor
-
-    
-    def searchColorInit(self):
-        self.colorSensor.colorSensor.setGain(rev.color._rev_color.ColorSensorV3.GainFactor.k18x)
-        self.currentColor = None
-        self.lastColor = None
-        self.goal = self.stopColorMap[self.gameData]
-        self.goal = self.stopColorMap[self.goal]
-        #self.goal = self.gameData
-        self.searchForColor = True
-        self.found = False
-        self.timer = 0
-        self.timer2 = 2
-        
-
-    def searchColorCycle(self):
-        if self.timer < 100:
-            self.timer += 1
-            self.colorSensorMotor.set(0.2)
-            
-        elif self.timer == 100:
-            self.currentColor = self.colorMatch.matchClosestColor(self.colorSensor.getWPIColor(), 1)
-            self.timer +=1
-
-        else:
-
-            self.lastColor = self.currentColor
-
-            color = self.colorMatch.matchClosestColor(self.colorSensor.getWPIColor(), 0.95)
-            
-            self.currentColor = color
-
-            if self.lastColor == self.yellow and self.currentColor == self.green:
-                self.currentColor == self.yellow
-
-            if self.colorMap[self.goal] != color:
-                self.colorSensorMotor.set(0.2)
-
-            else:
-
-                if self.timer2 == 0:
-                    print("STOP!!!!")    
-                    self.colorSensorMotor.set(0)
-                    self.searchForColor = False
-                else:
-                    self.timer2 -= 1
 
     def shiftGears(self):
         if self.driver.getBumperPressed(RIGHT_HAND):
             if self.gearshiftPosition == "Low":
                 self.gearshiftPiston.retract()
                 self.gearshiftPosition = "High"
-                #print("Shifted to high gear")
             else:
                 self.gearshiftPiston.extend()
                 self.gearshiftPosition = "Low"
-                #print("Shifted to low gear")
+
 
     def visionShooterUpdate(self):
         """
@@ -396,31 +222,21 @@ class Robot(wpilib.TimedRobot):
 
     def teleopPeriodic(self):
         forward = self.driver.getY(RIGHT_HAND) 
-        #Right stick y-axis
+
         forward = 0.80 * deadzone(forward, robotmap.deadzone)
         rotation_value = -0.8 * self.driver.getX(LEFT_HAND)
         
-        #if rotation_value > 0 or forward > 0:
         self.drivetrain.arcadeDrive(forward, rotation_value)
 
-        self.checkGameData()
+        self.colorSensor.checkGameData()
 
         if self.operator.getStartButtonPressed():
-            if not self.hasTurnedWheel:
-                self.turnWheelInit()
-            else:
-                self.searchColorInit()
+            self.colorSensor.colorWheelCycle()
             
         if self.operator.getBackButton():
             self.colorSensorMotor.set(0.2)
         elif not self.searchForColor and not self.turnWheel:
             self.colorSensorMotor.set(0)
-
-        if self.searchForColor:
-            self.searchColorCycle()
-
-        if self.turnWheel:
-            self.turnWheelCycle()
 
         self.colorPistonUpdate()
         self.climberPistonUpdate()
